# Fix Performance Regression caused by split-or