#2238 introduced a performance regression in TPC-DS:

```diff
-| 10      ||     19.7 |    6060.5 |   +30674%  ||    50.77 |     0.17 | -100%  |  0.0000 |
-| 35      ||     89.2 | 1303100.2 | +1461565%  ||    11.22 |     0.00 | -100%  |       ˅ |
-| 45      ||      9.8 |      17.4 |      +77%  ||   101.89 |    57.59 |  -43%  |  0.0000 |
```

This went unnoticed because <s>I</s> someone did not have sufficient compute resources and only ran the TPC-H benchmark. What happened was that some disjunctions were made unavailable for the SubqueryToJoinRule.

With this PR, that someone reverts the old order in the optimizer and instead fixes the ColumnPruningRule to be able to deal with certain types of UnionAll.

<details>
<summary>Here is the complete benchmark_all result for #2238</summary>

**System**
<details>
<summary>nemea - click to expand</summary>

| property | value |
| -- | -- |
| Hostname | nemea |
| CPU | Intel(R) Xeon(R) Platinum 8180 CPU @ 2.50GHz |
| Memory | 1.2TB |
| numactl | nodebind: 0  |
| numactl | membind: 0  |
</details>

**Commit Info and Build Time**
| commit | date | message | build time |
| -- | -- | -- | -- |
| 44f8c4aa8 | 08.10.2020 21:29 | Create empty bloom filter only once (#2236) | real 270.68 user 2492.41 sys 152.64|
| f518e3173 | 10.10.2020 21:19 | PredicateSplitUpRule: Create UnionAll for NOT LIKE (#2238) | real 269.13 user 2467.01 sys 148.58|


**hyriseBenchmarkTPCH - single-threaded**
<details>
<summary>
Sum of avg. item runtimes: -3%
 ||
Geometric mean of throughput changes: +3%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Markus.Dreseler/hyrise5/build-release/benchmark_all_results/hyriseBenchmarkTPCH_44f8c4aa8ffb1b6010213922c15aa04b551bc73c_st.json | /home/Markus.Dreseler/hyrise5/build-release/benchmark_all_results/hyriseBenchmarkTPCH_f518e3173f1aabef3983b361fc6aa7fa81190d51_st.json |
 +--------------------------+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | 44f8c4aa8ffb1b6010213922c15aa04b551bc73c-dirty                                                                                         | f518e3173f1aabef3983b361fc6aa7fa81190d51-dirty                                                                                         |
 |  benchmark_mode          | Ordered                                                                                                                                | Ordered                                                                                                                                |
 |  build_type              | release                                                                                                                                | release                                                                                                                                |
 |  chunk_size              | 65535                                                                                                                                  | 65535                                                                                                                                  |
 |  clients                 | 1                                                                                                                                      | 1                                                                                                                                      |
 |  compiler                | gcc 9.2                                                                                                                                | gcc 9.2                                                                                                                                |
 |  cores                   | 0                                                                                                                                      | 0                                                                                                                                      |
 |  date                    | 2020-10-12 09:35:29                                                                                                                    | 2020-10-13 22:34:03                                                                                                                    |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                | {'default': {'encoding': 'Dictionary'}}                                                                                                |
 |  indexes                 | False                                                                                                                                  | False                                                                                                                                  |
 |  max_duration            | 60000000000                                                                                                                            | 60000000000                                                                                                                            |
 |  max_runs                | -1                                                                                                                                     | -1                                                                                                                                     |
 |  scale_factor            | 1.0                                                                                                                                    | 1.0                                                                                                                                    |
 |  time_unit               | ns                                                                                                                                     | ns                                                                                                                                     |
 |  use_prepared_statements | False                                                                                                                                  | False                                                                                                                                  |
 |  using_scheduler         | False                                                                                                                                  | False                                                                                                                                  |
 |  verify                  | False                                                                                                                                  | False                                                                                                                                  |
 |  warmup_duration         | 0                                                                                                                                      | 0                                                                                                                                      |
 +--------------------------+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
```
</details>

```diff
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||    800.1 |   827.2 |   +3%  ||     1.25 |     1.21 |   -3%  |  0.0000 |
 | TPC-H 02 ||      5.0 |     5.0 |   -0%  ||   199.76 |   200.71 |   +0%  |  0.5311 |
 | TPC-H 03 ||     81.8 |    81.3 |   -1%  ||    12.22 |    12.31 |   +1%  |  0.0000 |
 | TPC-H 04 ||     68.7 |    68.4 |   -0%  ||    14.56 |    14.62 |   +0%  |  0.0000 |
+| TPC-H 05 ||    232.3 |   216.8 |   -7%  ||     4.30 |     4.61 |   +7%  |  0.0000 |
 | TPC-H 06 ||      3.6 |     3.6 |   -1%  ||   277.13 |   279.00 |   +1%  |  0.0000 |
 | TPC-H 07 ||     60.0 |    59.6 |   -1%  ||    16.67 |    16.78 |   +1%  |  0.5025 |
 | TPC-H 08 ||     69.3 |    69.4 |   +0%  ||    14.43 |    14.42 |   -0%  |  0.8406 |
 | TPC-H 09 ||    497.0 |   485.8 |   -2%  ||     2.01 |     2.06 |   +2%  |  0.0000 |
 | TPC-H 10 ||    168.9 |   169.2 |   +0%  ||     5.92 |     5.91 |   -0%  |  0.7903 |
 | TPC-H 11 ||      9.1 |     9.1 |   -1%  ||   109.45 |   110.12 |   +1%  |  0.0000 |
 | TPC-H 12 ||     43.8 |    43.6 |   -0%  ||    22.82 |    22.93 |   +0%  |  0.0001 |
+| TPC-H 13 ||    434.1 |   360.5 |  -17%  ||     2.30 |     2.77 |  +20%  |  0.0000 |
 | TPC-H 14 ||     21.3 |    21.0 |   -2%  ||    46.88 |    47.71 |   +2%  |  0.0000 |
 | TPC-H 15 ||      7.9 |     7.8 |   -0%  ||   126.76 |   127.35 |   +0%  |  0.0000 |
+| TPC-H 16 ||     95.3 |    75.4 |  -21%  ||    10.50 |    13.26 |  +26%  |  0.0000 |
 | TPC-H 17 ||     18.2 |    18.2 |   -0%  ||    54.94 |    55.04 |   +0%  |  0.3465 |
 | TPC-H 18 ||    708.7 |   682.0 |   -4%  ||     1.41 |     1.47 |   +4%  |  0.0010 |
 | TPC-H 19 ||     29.9 |    29.7 |   -1%  ||    33.43 |    33.67 |   +1%  |  0.0000 |
 | TPC-H 20 ||     16.3 |    16.3 |   -0%  ||    61.36 |    61.44 |   +0%  |  0.6615 |
 | TPC-H 21 ||    425.8 |   422.0 |   -1%  ||     2.35 |     2.37 |   +1%  |  0.0429 |
 | TPC-H 22 ||     50.6 |    49.4 |   -2%  ||    19.76 |    20.24 |   +2%  |  0.0000 |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Sum      ||   3847.8 |  3721.1 |   -3%  ||          |          |        |         |
 | Geomean  ||          |         |        ||          |          |   +3%  |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
```
</details>


**hyriseBenchmarkTPCH - single-threaded, SF 0.01**
<details>
<summary>
Sum of avg. item runtimes: -4%
 ||
Geometric mean of throughput changes: +4%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Markus.Dreseler/hyrise5/build-release/benchmark_all_results/hyriseBenchmarkTPCH_44f8c4aa8ffb1b6010213922c15aa04b551bc73c_st_s01.json | /home/Markus.Dreseler/hyrise5/build-release/benchmark_all_results/hyriseBenchmarkTPCH_f518e3173f1aabef3983b361fc6aa7fa81190d51_st_s01.json |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | 44f8c4aa8ffb1b6010213922c15aa04b551bc73c-dirty                                                                                             | f518e3173f1aabef3983b361fc6aa7fa81190d51-dirty                                                                                             |
 |  benchmark_mode          | Ordered                                                                                                                                    | Ordered                                                                                                                                    |
 |  build_type              | release                                                                                                                                    | release                                                                                                                                    |
 |  chunk_size              | 65535                                                                                                                                      | 65535                                                                                                                                      |
 |  clients                 | 1                                                                                                                                          | 1                                                                                                                                          |
 |  compiler                | gcc 9.2                                                                                                                                    | gcc 9.2                                                                                                                                    |
 |  cores                   | 0                                                                                                                                          | 0                                                                                                                                          |
 |  date                    | 2020-10-12 09:58:12                                                                                                                        | 2020-10-13 22:56:21                                                                                                                        |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                    | {'default': {'encoding': 'Dictionary'}}                                                                                                    |
 |  indexes                 | False                                                                                                                                      | False                                                                                                                                      |
 |  max_duration            | 60000000000                                                                                                                                | 60000000000                                                                                                                                |
 |  max_runs                | -1                                                                                                                                         | -1                                                                                                                                         |
 |  scale_factor            | 0.009999999776482582                                                                                                                       | 0.009999999776482582                                                                                                                       |
 |  time_unit               | ns                                                                                                                                         | ns                                                                                                                                         |
 |  use_prepared_statements | False                                                                                                                                      | False                                                                                                                                      |
 |  using_scheduler         | False                                                                                                                                      | False                                                                                                                                      |
 |  verify                  | False                                                                                                                                      | False                                                                                                                                      |
 |  warmup_duration         | 0                                                                                                                                          | 0                                                                                                                                          |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
```
</details>

```diff
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||      7.3 |     7.3 |   -0%  ||   136.40 |   137.00 |   +0%  |  0.0000 |
 | TPC-H 02 ||      0.5 |     0.5 |   -1%  ||  1836.58 |  1855.86 |   +1%  |  0.0167 |
 | TPC-H 03 ||      0.9 |     0.9 |   -2%  ||  1109.36 |  1136.81 |   +2%  |  0.0000 |
 | TPC-H 04 ||      0.5 |     0.5 |   -2%  ||  1907.08 |  1937.88 |   +2%  |  0.0000 |
 | TPC-H 05 ||      1.3 |     1.2 |   -4%  ||   789.40 |   819.03 |   +4%  |  0.0000 |
 | TPC-H 06 ||      0.1 |     0.1 |   -2%  ||  9181.73 |  9339.90 |   +2%  |  0.0000 |
 | TPC-H 07 ||      1.3 |     1.2 |   -4%  ||   777.75 |   810.40 |   +4%  |  0.0006 |
 | TPC-H 08 ||     15.3 |    14.7 |   -4%  ||    65.45 |    68.01 |   +4%  |  0.0000 |
 | TPC-H 09 ||      2.4 |     2.4 |   -3%  ||   411.65 |   425.20 |   +3%  |  0.0000 |
 | TPC-H 10 ||      0.9 |     0.9 |   -3%  ||  1079.99 |  1111.68 |   +3%  |  0.0000 |
 | TPC-H 11 ||      0.3 |     0.3 |   -2%  ||  3309.06 |  3366.76 |   +2%  |  0.0000 |
 | TPC-H 12 ||      0.6 |     0.6 |   -3%  ||  1549.18 |  1593.98 |   +3%  |  0.0000 |
+| TPC-H 13 ||      2.4 |     2.2 |   -9%  ||   414.33 |   455.73 |  +10%  |  0.0000 |
 | TPC-H 14 ||      0.4 |     0.3 |   -3%  ||  2773.34 |  2863.84 |   +3%  |  0.0000 |
 | TPC-H 15 ||      1.1 |     1.1 |   -4%  ||   868.63 |   901.56 |   +4%  |  0.0000 |
+| TPC-H 16 ||      2.2 |     2.0 |  -10%  ||   448.88 |   498.11 |  +11%  |  0.0000 |
…